Was working nicly before checked-in ; but because of modification-time changes (due to the check-in), dependencies have changed, and is now failling when trying to run Automake/Autoconf. => go back to previous version (hopefully should works).

+++ tools/mpack-1.6/md5.h
@@ -34,16 +34,14 @@ The following makes PROTOTYPES default to 0 if it has not already
 #define PROTOTYPES 0
 #endif
 
-#include <stdint.h>
-
 /* POINTER defines a generic pointer type */
 typedef unsigned char *POINTER;
 
 /* UINT2 defines a two byte word */
-typedef uint16_t UINT2;
+typedef unsigned short int UINT2;
 
 /* UINT4 defines a four byte word */
-typedef uint32_t UINT4;
+typedef unsigned long int UINT4;
 
 /* PROTO_LIST is defined depending on how PROTOTYPES is defined above.
 If using PROTOTYPES, then PROTO_LIST returns the list, otherwise it
